Completed species management.

+++ backend/src/PokeGame.Core/Moves/Models/MoveModel.cs
@@ -10,9 +10,9 @@ namespace PokeGame.Core.Moves.Models
     public string Name { get; set; } = null!;
     public string? Description { get; set; }
 
-    public double? Accuracy { get; set; }
-    public int? Power { get; set; }
-    public int PowerPoints { get; set; }
+    public byte? Accuracy { get; set; }
+    public byte? Power { get; set; }
+    public byte PowerPoints { get; set; }
 
     public string? Notes { get; set; }
     public string? Reference { get; set; }
